styles folder is better organized for classs styles for components and layouts only
