+++ src/App.jsx
@@ -1,14 +1,16 @@
 import './styles/index.css'
 import { Navbar } from './components/layout/Navbar'
 import { Footer } from './components/layout/Footer'
-import HereSection from './Pages/Home/hero-section'
+import HeroSection from './Pages/Home/heroSection'
+import LastestProjectsSection from './Pages/Home/lasterProjectsSection'
 function App() {
   
 
   return (
     <>
       <Navbar></Navbar>
-      <HereSection></HereSection>
+      <HeroSection></HeroSection>
+      <LastestProjectsSection></LastestProjectsSection>
       <Footer></Footer>
     </>
   )
