+++ src/Pages/Home/lasterProjectsSection.jsx
@@ -1,0 +1,9 @@
+import HeadLine from "../../components/common/headlines";
+
+export default function LastestProjectsSection(){
+    return (
+        <>
+            <HeadLine></HeadLine>
+        </>
+    );
+}
