+++ src/components/common/headlines.jsx
@@ -1,0 +1,6 @@
+
+export default function HeadLine(){
+    return (
+        <h1 className="text-2xl">xD</h1>
+    );
+}
