+++ src/components/layout/headlinesLayout.jsx
@@ -1,0 +1,7 @@
+export function HeadlineLayout() {
+    return (
+        <div>
+            
+        </div>
+    );
+}
