+++ src/styles/layouts/headlines-layout.js
@@ -1,0 +1,12 @@
+/* 
+    This file contains all of the styles
+    used in the main layouts following the tailwind css
+    style of using classes to style components
+    instead of pure css.
+*/
+
+const LAYOUTS_STYLES = {
+    headers: "flex items-center",
+};
+
+export default LAYOUTS_STYLES;
